+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project follows [Semantic Versioning](http://semver.org/).
 ## [1.5.0] - 2020-05-06
 ### Changed
 - Allowed searching of any searchable Patient field
+- Added display of data loading time
 
 ## [1.4.0] - 2020-04-23
 ### Changed